+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # YTech Changelog
 
+## []
+
+- Fixed not rendering water in Aqueduct
+
 ## [0.4.5]
 
 - Fixed Aqueduct Fertilizer not working